Update train page functionality
This updates the functionality of the train page to store a global
variable for denoting the player's current turn. This also updates the
space layout index to get the last known space layout index whenever an
analyzer instance is being loaded. This also updates the displaying of
die rolls a player can roll with any restrictions put in place for the
player.