Add player movement options
This adds styling and functionality capabilities for the player to move
around the board. This will allow the user to select a die and roll that
number of spaces to land on. This will also let the user choose a
direction to choose from when given multiple directions a player can go
to.