Update dependency injection context usage
This updates the dependency injection used for database contexts to be
done from the startup program file instead of the Entity Framework
context initialization. The database context used for saving analyzer
instance logs, a deprecated feature, has been updated to save post roll
data. This also fixes an issue with the authentication flow of
dependency injection for production environment in Azure.