Update animate functions
This updates functions used for the flashing animation to be condensed
into one function. This will allow multiple flashing animations to occur
at the same time. Also, the function will now stop the flashing
animation before the animation is complete whenever there is a change in
the state of the animation.